chore(site): replace PasswordField with non-MUI version
