+++ site/src/pages/UserSettingsPage/SecurityPage/SecurityForm.tsx
@@ -1,4 +1,3 @@
-import TextField from "@mui/material/TextField";
 import { type FormikContextType, useFormik } from "formik";
 import type { FC } from "react";
 import * as Yup from "yup";
@@ -6,6 +5,7 @@ import { Alert } from "#/components/Alert/Alert";
 import { ErrorAlert } from "#/components/Alert/ErrorAlert";
 import { Button } from "#/components/Button/Button";
 import { Form, FormFields } from "#/components/Form/Form";
+import { FormField } from "#/components/FormField/FormField";
 import { PasswordField } from "#/components/PasswordField/PasswordField";
 import { Spinner } from "#/components/Spinner/Spinner";
 import { getFormHelpers } from "#/utils/formUtils";
@@ -67,25 +67,22 @@ export const SecurityForm: FC<SecurityFormProps> = ({
 		<Form onSubmit={form.handleSubmit}>
 			<FormFields>
 				{Boolean(error) && <ErrorAlert error={error} />}
-				<TextField
-					{...getFieldHelpers("old_password")}
-					autoComplete="old_password"
-					fullWidth
+				<FormField
+					field={getFieldHelpers("old_password")}
 					label="Old Password"
 					type="password"
+					autoComplete="current-password"
 				/>
 				<PasswordField
-					{...getFieldHelpers("password")}
-					autoComplete="password"
-					fullWidth
+					field={getFieldHelpers("password")}
 					label="New Password"
+					autoComplete="new-password"
 				/>
-				<TextField
-					{...getFieldHelpers("confirm_password")}
-					autoComplete="confirm_password"
-					fullWidth
+				<FormField
+					field={getFieldHelpers("confirm_password")}
 					label="Confirm Password"
 					type="password"
+					autoComplete="new-password"
 				/>
 
 				<div>
